Reorganize ReadonlyDateWithNanoseconds
