+++ packages/tendermint-rpc/src/types.ts
@@ -1,5 +1,11 @@
 // Types in this file are exported outside of the @cosmjs/tendermint-rpc package,
 // e.g. as part of a request or response
+import { ReadonlyDate } from "readonly-date";
+
+export interface ReadonlyDateWithNanoseconds extends ReadonlyDate {
+  /* Nanoseconds after the time stored in a vanilla ReadonlyDate (millisecond granularity) */
+  readonly nanoseconds?: number;
+}
 
 export interface ValidatorEd25519Pubkey {
   readonly algorithm: "ed25519";
@@ -23,6 +29,6 @@ export enum BlockIdFlag {
 export interface CommitSignature {
   blockIdFlag: BlockIdFlag;
   validatorAddress: Uint8Array;
-  timestamp?: Date;
+  timestamp?: ReadonlyDateWithNanoseconds;
   signature: Uint8Array;
 }
